Add a table to display deleted-policies And Move class into mvc package to differenciate with eventsourcing code
