--- src/main/java/com/hodvidar/insuratradeflow/api/config/CorsConfig.java
@@ -1,16 +1,0 @@
-package com.hodvidar.insuratradeflow.api.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CorsConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("http://localhost:3000", "http://frontend:3000")
-                .allowedMethods("*")
-                .allowCredentials(true);
-    }
-}
