--- src/main/java/com/hodvidar/insuratradeflow/api/config/WebSecurityConfig.java
@@ -1,87 +1,0 @@
-package com.hodvidar.insuratradeflow.api.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-
-import static org.springframework.security.config.Customizer.withDefaults;
-
-@Configuration
-@EnableWebSecurity
-public class WebSecurityConfig {
-
-    private static final String[] AUTH_WHITELIST = {
-            // -- Swagger UI v2
-            "/v2/api-docs",
-            "v2/api-docs",
-            "/swagger-resources",
-            "swagger-resources",
-            "/swagger-resources/**",
-            "swagger-resources/**",
-            "/configuration/ui",
-            "configuration/ui",
-            "/configuration/security",
-            "configuration/security",
-            "/swagger-ui.html",
-            "swagger-ui.html",
-            "/webjars/**",
-            "webjars/**",
-            "/api/public/**",
-            "/api/public/authenticate",
-            "/swagger-ui/**",
-            // -- Swagger UI v3
-            "/v3/api-docs/**",
-            "v3/api-docs/**",
-            "/swagger-ui/**",
-            "swagger-ui/**",
-            // CSA Controllers
-            "/csa/api/token",
-            // Actuators
-            "/actuator/**",
-            "/health/**",
-            "/api/greeting" // our test endpoint
-    };
-
-    @Value("${security.enable-csrf}")
-    private boolean csrfEnabled;
-
-    @Value("${security.user.name}")
-    private String userName;
-    @Value("${security.user.password}")
-    private String userPassword;
-    @Value("${security.user.role}")
-    private String userRole;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests(request -> request
-                        .requestMatchers(AUTH_WHITELIST).permitAll()
-                        .anyRequest().authenticated()
-                )
-                .httpBasic(withDefaults());
-        if (!csrfEnabled) {
-            http.csrf(AbstractHttpConfigurer::disable);
-        }
-        return http.build();
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        UserDetails user =
-                User.withDefaultPasswordEncoder()
-                        .username(userName)
-                        .password(userPassword)
-                        .roles(userRole)
-                        .build();
-
-        return new InMemoryUserDetailsManager(user);
-    }
-}
