--- src/main/java/com/hodvidar/insuratradeflow/api/controller/ControllerExceptionHandler.java
@@ -1,58 +1,0 @@
-package com.hodvidar.insuratradeflow.api.controller;
-
-import com.hodvidar.insuratradeflow.business.validation.InsurancePolicyValidationException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-@Slf4j
-@RestControllerAdvice
-public class ControllerExceptionHandler {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, List<String>>> handleValidationErrors(MethodArgumentNotValidException ex) {
-        log.error("Validation error: ", ex);
-        return new ResponseEntity<>(getErrorsMap(getErrors(ex)),
-                new HttpHeaders(),
-                HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(InsurancePolicyValidationException.class)
-    public ResponseEntity<String> handleDataIntegrityViolationErrors(InsurancePolicyValidationException ex) {
-        log.error("Validation error: ", ex);
-        return new ResponseEntity<>(ex.getMessage(), new HttpHeaders(), HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<String> handleDataIntegrityViolationErrors(DataIntegrityViolationException ex) {
-        log.error("Data integrity violation: ", ex);
-        return new ResponseEntity<>(ex.getMessage(), new HttpHeaders(), HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> handleDataIntegrityViolationErrors(IllegalArgumentException ex) {
-        log.error("Data integrity violation: ", ex);
-        return new ResponseEntity<>(ex.getMessage(), new HttpHeaders(), HttpStatus.BAD_REQUEST);
-    }
-
-    private List<String> getErrors(final MethodArgumentNotValidException ex) {
-        return ex.getBindingResult().getFieldErrors()
-                .stream().map(FieldError::getDefaultMessage).toList();
-    }
-
-    private Map<String, List<String>> getErrorsMap(List<String> errors) {
-        Map<String, List<String>> errorResponse = new HashMap<>();
-        errorResponse.put("errors", errors);
-        return errorResponse;
-    }
-}
