--- src/main/java/com/hodvidar/insuratradeflow/api/dto/InsurancePolicyDto.java
@@ -1,15 +1,0 @@
-package com.hodvidar.insuratradeflow.api.dto;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-public record InsurancePolicyDto(
-        Long id,
-        String name,
-        String status, // ACTIVE or INACTIVE
-        LocalDate startDate,
-        LocalDate endDate,
-        LocalDateTime creationDateTime,
-        LocalDateTime updateDateTime
-) {
-}
