--- src/main/java/com/hodvidar/insuratradeflow/business/domain/PolicyStatus.java
@@ -1,6 +1,0 @@
-package com.hodvidar.insuratradeflow.business.domain;
-
-public enum PolicyStatus {
-    ACTIVE,
-    INACTIVE
-}
