--- src/main/java/com/hodvidar/insuratradeflow/business/validation/InsurancePolicyValidationException.java
@@ -1,8 +1,0 @@
-package com.hodvidar.insuratradeflow.business.validation;
-
-
-public class InsurancePolicyValidationException extends Exception {
-    public InsurancePolicyValidationException(String message) {
-        super(message);
-    }
-}
