+++ src/main/java/com/hodvidar/insuratradeflow/eventsourcing/command/DeleteInsurancePolicyCommand.java
@@ -1,6 +1,6 @@
 package com.hodvidar.insuratradeflow.eventsourcing.command;
 
-import com.hodvidar.insuratradeflow.business.domain.InsurancePolicy;
+import com.hodvidar.insuratradeflow.mvc.business.domain.InsurancePolicy;
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
 
 public record DeleteInsurancePolicyCommand(
