+++ src/main/java/com/hodvidar/insuratradeflow/eventsourcing/event/InsurancePolicyDeletedEvent.java
@@ -1,6 +1,6 @@
 package com.hodvidar.insuratradeflow.eventsourcing.event;
 
-import com.hodvidar.insuratradeflow.business.domain.InsurancePolicy;
+import com.hodvidar.insuratradeflow.mvc.business.domain.InsurancePolicy;
 
 public record InsurancePolicyDeletedEvent(
         Long policyId,
