+++ src/main/java/com/hodvidar/insuratradeflow/mvc/business/domain/PolicyStatus.java
@@ -1,0 +1,6 @@
+package com.hodvidar.insuratradeflow.mvc.business.domain;
+
+public enum PolicyStatus {
+    ACTIVE,
+    INACTIVE
+}
