+++ src/main/java/com/hodvidar/insuratradeflow/mvc/business/validation/InsurancePolicyValidationException.java
@@ -1,0 +1,8 @@
+package com.hodvidar.insuratradeflow.mvc.business.validation;
+
+
+public class InsurancePolicyValidationException extends Exception {
+    public InsurancePolicyValidationException(String message) {
+        super(message);
+    }
+}
