--- src/main/java/com/hodvidar/insuratradeflow/persistance/repository/InsurancePolicyRepository.java
@@ -1,16 +1,0 @@
-package com.hodvidar.insuratradeflow.persistance.repository;
-
-import com.hodvidar.insuratradeflow.persistance.dao.InsurancePolicyDao;
-import lombok.NonNull;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface InsurancePolicyRepository extends JpaRepository<InsurancePolicyDao, Long> {
-    @NonNull
-    Page<InsurancePolicyDao> findAll(@NonNull Pageable pageable);
-
-    // If we need filtered pagination later:
-    // Page<InsurancePolicyDao> findByNameContaining(String nameFilter, Pageable pageable);
-}
-
